fix(response): Moved APIRequester to separate file to reduce clutter
